linear pid implemented partially for depth

+++ motion_controller/src/trap_int.cpp
@@ -6,6 +6,6 @@ void TrapezoidalIntegrator::add(double dt, double value) {
 	prevPoint=value;
 }
 
-float TrapezoidalIntegrator::get() const {
+double TrapezoidalIntegrator::get() const {
 	return total;
 }
